Apply different checkers for order pre and post completion
